camera_recorder/requirements.txt
opencv-python>=4.5.0
pyyaml>=6.0
numpy>=1.21.0
